+++ CHANGELOG.md
@@ -19,6 +19,7 @@ delimited CGI params by default.
 
 ### Bug Fixes
 
+- Fix numeric instability with `pyiem.plot.centered_bins` (#871).
 - Forward propagate sbw database storage of VTEC issuace (#862).
 - Improve NWS Text Product signature logic so to more generally match what
 looks like a signature (#865).